perf(e2e): shard the suite into three functional groups
The E2E job ran all 17 suites in one sequential mocha process. Splitting it
across a matrix cuts wall time, but a split has to be maintainable: a list of
files balanced by measured duration goes stale the moment someone adds a suite,
and the way it goes stale is invisible.

Grouped by what the suites cover instead, one directory per shard:

  kernel/     4  running code and environments
  files/      6  the .deepnote file lifecycle — split, create, rename, delete
  workspace/  7  surfacing notebooks and reacting to changes on disk

A new suite has an obvious home, and the weights happen to land close enough
(~150s / ~100s / ~110s against the post-venv-share profile) that no shard is
much more than 1.5x another. Balance is a consequence of the grouping rather
than something to re-tune.

Each shard pays the ~2m08s of setup again, which is what caps the return: three
shards is where the ratio peaks, and past four the setup dominates.

check:e2e:groups guards the failure mode that matters. A suite in the wrong place
does not fail anything — it silently never runs, which looks like a faster green
build. The check fails when a suite sits outside a group directory, when a group
has no matching npm script, when a group is not named in the workflow, and when a
script has no directory. Verified it exits 1 on a stray suite and 0 once moved
back.

Screenshot artifacts are per shard, since three jobs uploading one name collide.

Verified: tsc exits 0 after the move, the compiled tree matches the shard globs
(4/6/7), the workflow parses with the expected matrix, and the guard passes. NOT
verified: no E2E run yet.
