+++ test/e2e/suite/files/splitInitNotebook.e2e.test.ts
@@ -1,6 +1,6 @@
 /**
- * E2E (ExTester): the on-open split of the multi-notebook `sales-analytics.deepnote` (Overview,
- * Revenue, Forecast + a project BigQuery integration) into one single-notebook file per notebook.
+ * E2E (ExTester): splitting the multi-notebook `etl-pipeline.deepnote` (init "Init" + mains "Extract"/
+ * "Transform") emits the init notebook as its own sibling; every main keeps referencing it via initNotebookId.
  */
 
 import { expect } from 'chai';
@@ -20,9 +20,9 @@ import {
     assertNotNull,
     showView,
     waitForNotification
-} from '../helpers';
+} from '../../helpers';
 
-const FIXTURE = 'sales-analytics.deepnote';
+const FIXTURE = 'etl-pipeline.deepnote';
 const SPLIT_ACTION = 'Split into separate files';
 const SPLIT_PROMPT = /multiple notebooks/i;
 const SPLIT_DONE = /Split into \d+ files\./i;
@@ -30,29 +30,19 @@ const SPLIT_DONE = /Split into \d+ files\./i;
 // How long to wait while confirming a single-notebook file does NOT raise the split prompt.
 const NO_SPLIT_PROMPT_TIMEOUT = 6_000;
 
-const INTEGRATION_NAME = 'Sales BigQuery';
+// The fixture's `project.initNotebookId`; the init sibling's notebook id must equal this.
+const INIT_NOTEBOOK_ID = 'b-nb-init';
 
-// The three single-notebook files the split should produce; each marker is unique to its source
-// notebook, proving the right blocks landed in the right file.
-const SIBLINGS: ReadonlyArray<{ file: string; notebookName: string; contentMarkers: string[] }> = [
-    {
-        file: 'sales-analytics-overview.deepnote',
-        notebookName: 'Overview',
-        contentMarkers: ['type: text-cell-h1', 'print("Overview notebook")']
-    },
-    {
-        file: 'sales-analytics-revenue.deepnote',
-        notebookName: 'Revenue',
-        contentMarkers: ['type: sql', 'SELECT region, SUM(amount)']
-    },
-    {
-        file: 'sales-analytics-forecast.deepnote',
-        notebookName: 'Forecast',
-        contentMarkers: ['print("Forecast notebook")']
-    }
+const INIT_SIBLING = { file: 'etl-pipeline-init.deepnote', notebookName: 'Init' };
+
+const MAIN_SIBLINGS: ReadonlyArray<{ file: string; notebookName: string }> = [
+    { file: 'etl-pipeline-extract.deepnote', notebookName: 'Extract' },
+    { file: 'etl-pipeline-transform.deepnote', notebookName: 'Transform' }
 ];
 
-describe('Deepnote — splitting a legacy multi-notebook .deepnote file into single-notebook files', function () {
+const ALL_SIBLINGS = [INIT_SIBLING, ...MAIN_SIBLINGS];
+
+describe('Deepnote — splitting a legacy multi-notebook .deepnote file that has an init notebook', function () {
     this.timeout(SUITE_TIMEOUT);
 
     let cleanupTempDir: (() => void) | undefined;
@@ -73,11 +63,11 @@ describe('Deepnote — splitting a legacy multi-notebook .deepnote file into sin
         await VSBrowser.instance.waitForWorkbench(WORKBENCH_TIMEOUT);
 
         // Open the workspace folder FIRST: the serializer reads snapshots relative to it, and
-        // without one deserialization blocks on a warning that never resolves headlessly.
+        // without one deserialization blocks headlessly.
         await openFolderViaDialog(tempDir);
         await VSBrowser.instance.waitForWorkbench(WORKBENCH_TIMEOUT);
 
-        await showView('Deepnote', '[split]');
+        await showView('Deepnote', '[split-init]');
         await screenshot('before-open-explorer');
 
         // Opening the multi-notebook file fires the splitter on the open event, raising the prompt.
@@ -99,8 +89,6 @@ describe('Deepnote — splitting a legacy multi-notebook .deepnote file into sin
 
         await prompt.takeAction(SPLIT_ACTION);
 
-        // The original is retired by renaming to `<name>.deepnote.legacy` (deterministic, no OS trash
-        // needed), and only after every child is written; "Failed to split" is a filesystem-error net.
         const outcome = assertNotNull(
             await waitForNotification(/Split into \d+ files\.|Failed to split/i, WORKBENCH_TIMEOUT, true),
             'split outcome notification'
@@ -112,26 +100,26 @@ describe('Deepnote — splitting a legacy multi-notebook .deepnote file into sin
         originalRemoved = !fs.existsSync(path.join(tempDir, FIXTURE));
         legacyBackupExists = fs.existsSync(path.join(tempDir, `${FIXTURE}.legacy`));
 
-        await showView('Deepnote', '[split]');
+        await showView('Deepnote', '[split-init]');
         await driver.sleep(1500);
         await screenshot('explorer-siblings');
 
-        await showView('Explorer', '[split]');
+        await showView('Explorer', '[split-init]');
         await driver.sleep(1500);
         await screenshot('file-explorer');
     });
 
     after(async function () {
         await new WebView().switchBack().catch((error) => {
-            console.warn('[split] switch back from webview during cleanup:', error);
+            console.warn('[split-init] switch back from webview during cleanup:', error);
         });
         await new EditorView().closeAllEditors().catch((error) => {
-            console.warn('[split] close all editors during cleanup:', error);
+            console.warn('[split-init] close all editors during cleanup:', error);
         });
         try {
             cleanupTempDir?.();
         } catch (error) {
-            console.warn('[split] remove temp workspace dir during cleanup:', error);
+            console.warn('[split-init] remove temp workspace dir during cleanup:', error);
         }
     });
 
@@ -140,19 +128,36 @@ describe('Deepnote — splitting a legacy multi-notebook .deepnote file into sin
         expect(promptMessage).to.contain('one file per notebook');
     });
 
-    it('splits into one single-notebook file per notebook and retires the original as a .legacy backup', function () {
+    it('splits into one file per notebook (init + mains) and retires the original as a .legacy backup', function () {
         expect(outcomeMessage, 'split outcome notification').to.match(SPLIT_DONE);
         expect(originalRemoved, 'original .deepnote no longer present').to.equal(true);
         expect(legacyBackupExists, 'original retained as a .deepnote.legacy backup').to.equal(true);
 
-        for (const sibling of SIBLINGS) {
+        for (const sibling of ALL_SIBLINGS) {
             const siblingPath = path.join(tempDir, sibling.file);
             expect(fs.existsSync(siblingPath), `${sibling.file} should exist`).to.equal(true);
+            expect(notebookCount(fs.readFileSync(siblingPath, 'utf8')), `${sibling.file} is single-notebook`).to.equal(
+                1
+            );
+        }
+    });
 
-            const yaml = fs.readFileSync(siblingPath, 'utf8');
-            expect(notebookCount(yaml), `${sibling.file} should contain exactly one notebook`).to.equal(1);
-            expect(yaml, `${sibling.file} should hold the "${sibling.notebookName}" notebook`).to.contain(
+    it('gives the init notebook its own single-notebook sibling file', function () {
+        const yaml = fs.readFileSync(path.join(tempDir, INIT_SIBLING.file), 'utf8');
+
+        expect(yaml, 'init sibling holds the "Init" notebook').to.contain(`name: ${INIT_SIBLING.notebookName}`);
+        expect(yaml, 'init sibling notebook id === initNotebookId').to.contain(`id: ${INIT_NOTEBOOK_ID}`);
+        expect(yaml, 'init sibling declares the same initNotebookId').to.contain(`initNotebookId: ${INIT_NOTEBOOK_ID}`);
+    });
+
+    it('keeps each main sibling referencing the init notebook via initNotebookId', function () {
+        for (const sibling of MAIN_SIBLINGS) {
+            const yaml = fs.readFileSync(path.join(tempDir, sibling.file), 'utf8');
+            expect(yaml, `${sibling.file} holds the "${sibling.notebookName}" notebook`).to.contain(
                 `name: ${sibling.notebookName}`
+            );
+            expect(yaml, `${sibling.file} still references the init notebook`).to.contain(
+                `initNotebookId: ${INIT_NOTEBOOK_ID}`
             );
         }
     });
@@ -160,7 +165,7 @@ describe('Deepnote — splitting a legacy multi-notebook .deepnote file into sin
     it('opens each split file directly, without prompting to split again', async function () {
         const driver = VSBrowser.instance.driver;
 
-        for (const sibling of SIBLINGS) {
+        for (const sibling of ALL_SIBLINGS) {
             await dismissAllNotifications();
             await openWorkspaceFile(sibling.file);
             await driver.wait(
@@ -172,25 +177,6 @@ describe('Deepnote — splitting a legacy multi-notebook .deepnote file into sin
 
             const reprompt = await waitForNotification(SPLIT_PROMPT, NO_SPLIT_PROMPT_TIMEOUT, false);
             expect(reprompt, `${sibling.file} is single-notebook and must not prompt to split`).to.equal(undefined);
-        }
-    });
-
-    it('preserves each source notebook’s content in its own split file', function () {
-        for (const sibling of SIBLINGS) {
-            const yaml = fs.readFileSync(path.join(tempDir, sibling.file), 'utf8');
-
-            for (const marker of sibling.contentMarkers) {
-                expect(yaml, `${sibling.file} should contain "${marker}"`).to.contain(marker);
-            }
-        }
-    });
-
-    it('carries the project integration into every split file', function () {
-        for (const sibling of SIBLINGS) {
-            const yaml = fs.readFileSync(path.join(tempDir, sibling.file), 'utf8');
-            expect(yaml, `${sibling.file} should keep the "${INTEGRATION_NAME}" integration`).to.contain(
-                `name: ${INTEGRATION_NAME}`
-            );
         }
     });
 });
