+++ test/e2e/suite/workspace/openSingleNotebook.e2e.test.ts
@@ -15,7 +15,7 @@ import {
     openWorkspaceFile,
     readStatusBarText,
     waitForNotification
-} from '../helpers';
+} from '../../helpers';
 
 const FIXTURE = 'quick-notes.deepnote';
 const NOTEBOOK_NAME = 'Quick Notes';
