+++ test/e2e/suite/workspace/revealInExplorer.e2e.test.ts
@@ -16,7 +16,7 @@ import {
     openFolderViaDialog,
     openWorkspaceFile,
     waitForNotification
-} from '../helpers';
+} from '../../helpers';
 
 const FIXTURE = 'quick-notes.deepnote';
 const NOTEBOOK_NAME = 'Quick Notes';
