+++ test/e2e/suite/workspace/statusBar.e2e.test.ts
@@ -16,7 +16,7 @@ import {
     openFolderViaDialog,
     openWorkspaceFile,
     waitForNotification
-} from '../helpers';
+} from '../../helpers';
 
 const FIXTURE = 'quick-notes.deepnote';
 const SCRATCH_FILE = 'clipboard-scratch.txt';
